+++ Assets/Scripts/Game/Action/Internal/Pending/Piece/DiurnalPending.cs
@@ -19,8 +19,8 @@ namespace Game.Action.Internal.Pending.Piece
 
         public DiurnalPending(int maker, int target) : base(maker)
         {
-            Maker = (ushort)maker;
-            Target = (ushort)target;
+            Maker = maker;
+            Target = target;
         }
 
         protected override void CompleteAction()
