+++ Assets/Scripts/Game/Action/Internal/Pending/Piece/MegalodonActivePending.cs
@@ -24,8 +24,8 @@ namespace Game.Action.Internal.Pending.Piece
 
         public MegalodonActivePending(int maker, int target) : base(maker)
         {
-            Maker = (ushort)maker;
-            Target = (ushort)target;
+            Maker = maker;
+            Target = target;
         }
 
         protected override void CompleteAction()
