+++ Assets/Scripts/Game/Action/Internal/Pending/Relic/ReliquaryPending.cs
@@ -13,8 +13,8 @@ namespace Game.Action.Internal.Pending.Relic
         public ReliquaryPending(Reliquary cp, int maker) : base(maker)
         {
             _reliquary = cp;
-            Target = (ushort)maker;
-            Maker = (ushort)maker;
+            Target = maker;
+            Maker = maker;
         }
 
         protected override void CompleteAction()
