+++ Assets/Scripts/Game/Managers/MatchManager.cs
@@ -45,7 +45,7 @@ namespace Game.Managers
             var config = new List<PieceConfig>(lineup.WhiteConfig);
             config.AddRange(lineup.BlackConfig);
             
-            foreach (var pieceConfig in config.Select(cfg => new PieceConfig(cfg.Type, cfg.Color, (ushort) PosMap(cfg.Index, StartingSize), cfg.AugmentationNames)))
+            foreach (var pieceConfig in config.Select(cfg => new PieceConfig(cfg.Type, cfg.Color,  PosMap(cfg.Index, StartingSize), cfg.AugmentationNames)))
             {
                 ActionManager.ExecuteImmediately(new SpawnPiece(pieceConfig));
             }
