+++ Assets/Scripts/Game/Save/Army/Troop.cs
@@ -8,18 +8,21 @@ using Game.Managers;
 namespace Game.Save.Army
 {
     [MemoryPackable]
-    public partial struct Troop: IComparable<Troop>
+    public partial struct Troop : IComparable<Troop>
     {
         public readonly string PieceType;
-        public readonly ushort Rank;
-        public readonly ushort File;
+        public readonly int Rank;
+
+        public readonly int File;
+
         // public readonly bool Side;
         public Dictionary<AugmentationSlot, AugmentationName> equippedAugmentation;
+
         public Troop(string pieceType, int rank, int file)
         {
             PieceType = pieceType;
-            Rank = (ushort) rank;
-            File = (ushort)file;
+            Rank = rank;
+            File = file;
             // Side = side;
             equippedAugmentation = new();
         }
@@ -31,10 +34,12 @@ namespace Game.Save.Army
             if (File < other.File) return -1;
             return File > other.File ? 1 : 0;
         }
+
         public PieceInfo GetPieceInfo()
         {
             return AssetManager.Ins.PieceData[PieceType];
         }
+
         public void EquipAugmentation(AugmentationName aug)
         {
             var slot = AssetManager.Ins.AugmentationData[aug].Slot;
