+++ Assets/Scripts/UX/UI/Ingame/ChrysosShop/ChrysosShop.cs
@@ -76,7 +76,7 @@ namespace UX.UI.Ingame.ChrysosShop
         public void Buy(string type)
         {
             PendingAction.CommitResult(new ChrysosUpgrade(_chrysos.Pos,
-                new PieceConfig(type, _chrysos.Color, (ushort)PendingAction.Target), _cost));
+                new PieceConfig(type, _chrysos.Color, PendingAction.Target), _cost));
             Disable();
         }
